Voltando commit

+++ src/pages/inventory/inventory.ts
@@ -6,6 +6,7 @@ import { Quest } from '../../models/quest';
 import { PopOverHomePage } from '../pop-over-home/pop-over-home';
 import { RewardServiceProvider } from '../../providers/reward-service/reward-service';
 import { QuestServiceProvider } from '../../providers/quest-service/quest-service';
+import { AccountServiceProvider } from '../../providers/account-service/account-service';
 
 /**
  * Generated class for the InventoryPage page.
@@ -34,14 +35,21 @@ export class InventoryPage {
     public utils: UtilsServiceProvider,
     public popoverCtrl: PopoverController,
     public rewardService: RewardServiceProvider,
-    public questService: QuestServiceProvider) {
+    public questService: QuestServiceProvider,
+    public accountService: AccountServiceProvider) {
   }
 
   ionViewDidLoad() {
     this.utils.loadingShow();
-    this.getReward();
-    this.getQuests();
-    this.utils.loadingHide();
+    let connectionState = this.accountService.connectionState();
+    if(connectionState){
+      this.getReward();
+      this.getQuests();
+      this.utils.loadingHide();
+    }else{
+      this.utils.loadingHide();
+      this.utils.creatToast('Verifique sua conexão com a internet para prosseguir!');
+    }
   }
 
   async getReward(){
@@ -71,16 +79,23 @@ export class InventoryPage {
 
   removeReward(reward:RecompUser){
     this.utils.loadingShow();
+    let connectionState = this.accountService.connectionState();
+    if(connectionState){
+      this.rewardService.removeReward(reward).then(() => {
+        this.utils.loadingHide();
+        this.utils.creatSimpleAlert('Recompensa consumida com sucesso');
+        this.getReward();
+        
+      }).catch((error: Error) => {
+        this.utils.loadingHide();
+        this.utils.creatSimpleAlert('Erro ao consumir recompensa');
+      });
+    }else{
+      this.utils.loadingHide();
+      this.utils.creatToast('Verifique sua conexão com a internet para prosseguir!');
+    }
 
-    this.rewardService.removeReward(reward).then(() => {
-      this.utils.loadingHide();
-      this.utils.creatSimpleAlert('Recompensa removida com sucesso');
-      this.getReward();
-      
-    }).catch((error: Error) => {
-      this.utils.loadingHide();
-      this.utils.creatSimpleAlert('Erro ao excluir recompensa');
-    });
+    
 
   }
